Mona Alkhammash
HOMEWORK_03

on a scale from 1 to 5, how comfortable were you with this assignment?
3
What was a win you had with this assignment?
Grid style
What was a challenge you had with this assignment?
Layout
Is there anything that you'd like some further information on?
No
Other stuff on your mind? work with anyone? tell us about it!
Thanks for your effort